refactor(perf): split /founder-playbook + /resources into server shell + motion island
35th + 36th in the pattern series.
